refactor: 불필요한 주석 제거 및 quiz 관련 클래스 리팩토링

+++ src/main/java/com/quiz/dto/QuizRequestDto.java
@@ -19,8 +19,6 @@ public class QuizRequestDto {
     @Size(max = 2000, message = "Java 코드 입력은 2000자 이하여야 합니다.")
     private String javaContent;
     @NotEmpty(message = "JavaScript 코드 입력은 필수입니다.")
-//    @Size(max = 2000, message = "JavaScript 코드 입력은 2000자 이하여야 합니다.")
-//    private String javaScriptContent;
 
     @NotEmpty(message = "1번 보기는 필수입니다.")
     private String first;
@@ -30,11 +28,9 @@ public class QuizRequestDto {
     private String third;
     @NotEmpty(message = "4번 보기는 필수입니다.")
     private String fourth;
-//    @NotEmpty(message = "5번 보기는 필수입니다.")
-//    private String fifth;
 
     @Range(min=1, max=10, message="Quiz 번호는 1에서 10 사이입니다.")
-    private Integer quizNumber;
+    private Long quizNumber;
     @Range(min=1, max=4, message="Quiz 정답은 1번에서 4번 사이입니다.")
-    private Integer quizAnswer;
+    private int quizAnswer;
 }
